requirements.txt
cmake
dlib
face_recognition
numpy
opencv-python
python-slugify
bs4
lxml
python3-flask